Edited lab 10.2

+++ lab10_1/lab10_1/Source.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 
-long long int pow(int a, int b)
+double pow(int a, int b)
 {
-    long long int result = 1;
+    double result = 1;
     for(int i = 0;i < b;i++)
     {
         result*=a;
@@ -16,7 +16,7 @@ int main()
     printf_s("Enter number: ");
     scanf_s("%d", &number);
 
-    long long int result = 0;
+    double result = 0;
     for(int i = 0;i < number;i++)
     {
         for(int ia = 0;ia < number;ia++)
@@ -24,5 +24,5 @@ int main()
             result += pow(i+1,ia+1);
         }
     }
-    printf_s("Result: %lld\n", result);
+    printf_s("Result: %lf\n", result);
 }
